token生成をserviceファイルに切り出す
