+++ app/javascript/services/token.ts
@@ -1,0 +1,11 @@
+export const getToken = () => {
+  const metaTag = document.querySelector("meta[name='csrf-token']");
+  let token: string;
+
+  if (!metaTag) {
+    throw new Error('CSRF token not found: ページを再読み込みしてください');
+  } else {
+    token = metaTag.getAttribute('content') || '';
+  }
+  return token;
+};
